Refactor terminal representation again
Remove Terminal trait in favour of a TerminalCapabilities struct, and
explicitly pattern match over supported capabilities while rendering.

This removes boxing and dynamic dispatch, and makes different features
and internal behaviour like HTTP access more explicit.  We now also
explicitly check what and how certain features are supported while
rendering; this gets rids of the NotSupportedError thing and the kludge
to handle it in a special way which I never quite liked.

Also simply how we check for access to resources:  Remove the resource
enum and instead use Url directly.  Via the scheme an URL already tells
us whether it’s local or not, so we just rely on the scheme and file
path conversion now.  This makes resolving references in markdown files
and checking access much simpler.