Update to syntect 4.1 and use fancy-regex
We used to pin syntect because its onig dependency wanted Clang and
bindgen to build in later versions.

However as of 4.0 syntect can alternatively use a pure regex engine
which is considerably easier to build.

We can now upgrade to syntect 4.1 and enable this pure regex engine to
simply the build.