Fixed SIGPIPE when piping large data in magic
The method write_all fail with "Broken pipe (os error 32)"
when the data being piped is large enough to overflow the
kernel pipe buffer and it terminates the process due to
SIGPIPE. Rust doesn't ignore that signal.

To workaround it we extract only the first 4kb of
interesting bytes that's sufficient for detecting the
mime type.

It fixes the related issue https://github.com/lunaryorn/mdcat/issues/124.