Fix SIGPIPE when piping large data in magic.rs
The method `write_all` fail with "Broken pipe (os error 32)" when the data being piped is large enough to be terminated due to SIGPIPE as the file command reads maximum 1048576 bytes from the data. See MAGIC_PARAM_BYTES_MAX in manpage libmagic(3). To workaround we extract only the first 1mb of interesting bytes that's sufficient for determining the mime-type of the image.