Create regression tests from commonmark spec
Add a little python script which parses the JSON database of test cases
from https://spec.commonmark.org/0.29/spec.json and writes the markdown
into a file.

The new spec test target generates a test for each of these markdown
files.  Each test tries to read a file containing expect output; if the
file exists write the actual output to another file for inspection and
compare it against the expected output.

If the file with expected output is missing create it and panic to fail
the test, to quickly generate updated output if the test case database
is updated.  Git tells which expections changed or are new, so that we
can take a look.

The prime purpose of these tests isn't to check that rendering is
"right" but to catch unexpected panics and errors while rendering and
regressions when changing the output format.

Closes GH-137