Query window pixel size on Unix
mdcat needs the window size in pixels to correctly draw images in some
terminals.  Unfortunately term_size doesn't expose ws_xpixel/ws_ypixel
from the winsize struct so we now do the ioctl ourselves.

This effectively gets rid of term_size on Unix system which has the
added benefit that we can now freely choose which device to query
instead of being limited to stdin, stdout and stderr.

We use this opportunity to query the controlling terminal directly, via
ctermid.  This implies that mdcat correctly detects the terminal size
even when all standard streams are redirected.