Add ureq feature, enabled by default.
ureq depends on rustls, which depends on ring, which runs on a limited
set of platforms. By making it possible to disable the ureq feature, we
allow mdcat to be built for platforms not supported by ring, while
maintaining most of its functionality.

Add CI run to test build with ureq disabled.