Generate completions per subcommand
We used to generate completions for the top-level command, but that's
not correct in a multicall binary: It includes the subcommands as
completable arguments, whereas we actually need to split completions per
subcommand.

Hence, iterate over the two relevant subcommands aka binary names, and
separately generate the completion for each corresponding subcommand.

This corrects completions for mdcat, and as a bonus we now also generate
completions for mdless which we didn't have before.

Closes GH-214