Do not sniff mime types
Instead rely on the Content-Type header (remote resources) and the file
extension (local resources) to identify SVG images.  This removes the
dependency on the file tool and on libmagic, and also the entire magic
module, which simplifies the code and the build a bit, and removes a
runtime dependency.

It also effectively enables SVG images on Windows which so far did
mostly not work because Windows neither includes the file tool nor
libmagic, so users had to go out of their way and install file/libmagic
to get SVG images support on Windows.