Refactor kitty image data definition
Write chunk header statically instead of filling and clearing a vector;
the latter was copied verbatim from the Python reference implementation
but I find it cumbersome and inelegant in Rust.

Merge KittyImage and KittyImageFormat into a single KittyImageData enum
whose type ensures that `size` is specified when it must be specified
according to the specification.  The original implementation did not
prevent writing an RGB image without size/height which is invalid
according to the specification.