Only use gethostname on Windows
On Unix we already have rustix in the tree anyway (e.g. via tempfile
which in turn comes from native-tls, etc.), and we already use it
ourselves for terminal size information, so we might just as well also
use it to get the hostname.

Not that gethostname() is a particularly involved dependency but
still...