Ignore all descriptions for rendered images
CommonMark allows all inline elements in the image description part[1],
not only plain text. So in this commit we change to ignore all such
elements instead of panicking.

To correctly handle nested images inside the image description, here we
push dummy 'RenderedImage' states to the stack. An alternative solution
would be maintaining a counter for nested image level and checking the
counter when dealing with end of image events, but that feels less
tasteful to me.

Note that this is a bit more lenient than the spec because it allows
block elements as well. However, this behaviour aligns with the upstream
library[2], and the more correct version is a bit tedious to implement
because we have to list all inline tags ourselves.

Add tests for markups in image descriptions

Those cases are actually covered in the 'commonmark-spec' test suite.
However, those tests use non-existent images so they are not checked for
rendered images. As such, in this commit we add a new case in
'samples/images.md', primarily to test markups in descriptions of
rendered images.

Closes GH-285.

[1] https://spec.commonmark.org/0.31.2/#images
[2] https://docs.rs/pulldown-cmark/0.9.6/src/pulldown_cmark/html.rs.html#377-399