Add limited support for tables
With this commit, mdcat can render basic tables. However, the support is
incomplete. E.g., inline markups are stripped, overlong texts are not
wrapped, and block attributes are mostly ignored.

We intend to support those in the future. So we chose to not use a TTY
table rendering library because it would be hard to plug the library
into our existing renderer.