Handle reference links again
Add a new data field to track pending links.  Whenever we do not write
an inline image or inline link we push the pending link to store the
destination URL and title; in the corresponding end event we pop the
pending link and add it as a new link reference definition.