Render task list marker in any inline state
Previously, we restricted task list markers to inline list items,
however we do no actually care for the kind of inline text, as our
marker is just one checkbox that we can render anywhere.

As a result, mdcat no longer crashes when rendering a task list with
nested paragraphs.

Fixes #302